Add test and fix for dGeomBoxPointDepth
This tests dGeomBoxPointDepth with a cube and
points inside and outside the box.

The calculation for points outside the box was incorrect,
returning the negative of the distance to the farthest
side of the box. I've change it to correctly compute
the distance to a face, edge, or corner of the box.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>